requirements.txt
fastapi
uvicorn
pydantic
flask
twilio
openai
python-dotenv
pymongo
requests
pandas
numpy
scrapy
beautifulsoup4
lxml
selenium
webdriver-manager
langchain
langchain-core
langchain-openai
langchain-community
langchain-chroma
langchain-text-splitters
chromadb
sentence-transformers
mcp
fastmcp
gradio



